refactor: MDXButton to Button in mdx
