+++ components/button/MDXButton.tsx
@@ -1,0 +1,16 @@
+import ButtonLink from "./ButtonLink"
+
+export const MDXButton = ({ children, href, ...props }: any) => {
+  return (
+    <ButtonLink
+      href={href}
+      variant="primary_filled"
+      size="lg"
+      className="my-4"
+      external
+      {...props}
+    >
+      {children}
+    </ButtonLink>
+  )
+}
