+++ content/about/contact.mdx
@@ -12,30 +12,30 @@ The Center for Computation and Visualization (CCV) is a center within the Univer
         <StyledCard title="Email" iconName="FaEnvelopeOpenText">
             Inquire about resources, support, or a potential collaboration for a research project.
 
-            <MDXButton href="mailto:support@ccv.brown.edu">
+            <Button href="mailto:support@ccv.brown.edu">
                 Email Us
-            </MDXButton>
+            </Button>
         </StyledCard>
 
         <StyledCard title="Submit a Ticket" iconName="FaTicketAlt">
             Open a ticket with our user services team. We will respond as soon as possible.
 
-            <MDXButton href="mailto:support@ccv.brown.edu">
+            <Button href="mailto:support@ccv.brown.edu">
                 Submit a Ticket
-            </MDXButton>
+            </Button>
         </StyledCard>
 
         <StyledCard title="CCV Slack" iconName="FaSlack">
             CCV Share is a collection of Slack channels where CCV Staff and the community will be available to discuss your
             questions.
 
-            <MDXButton
+            <Button
                 href="https://join.slack.com/t/ccv-share/shared_invite/enQtODY5OTQ3MTk0ODU1LTM4OWQyZjVlYWRmY2QxNWEyZjQ0NzEwMmRlNTRlZjYyMjM1Y2U5MDU1ZGFmMmRhZWIzNjliYmQzYTBiMzY2NzU">
                 Join the Slack
-            </MDXButton>
-            <MDXButton href="https://ithelp.brown.edu/kb/articles/terms-of-use-for-brown-enterprise-slack">
+            </Button>
+            <Button href="https://ithelp.brown.edu/kb/articles/terms-of-use-for-brown-enterprise-slack">
                 Terms of Services
-            </MDXButton>
+            </Button>
         </StyledCard>
     </CardGroup>
     </ContentSection>
@@ -54,9 +54,9 @@ The Computational Biology Core (CBC) staff are available at office hours to prov
 
             **WEDNESDAYS**: 3pm - 4pm
 
-            <MDXButton href="https://events.brown.edu/ccv/week">
+            <Button href="https://events.brown.edu/ccv/week">
                 Find Zoom Link
-            </MDXButton>
+            </Button>
         </StyledCard>
 
         <StyledCard title="Center for Computation and Visualization">
@@ -64,9 +64,9 @@ The Computational Biology Core (CBC) staff are available at office hours to prov
 
             **FRIDAYS**: 10am - 12pm
 
-            <MDXButton href="https://events.brown.edu/ccv/week">
+            <Button href="https://events.brown.edu/ccv/week">
                 Find Zoom Link
-            </MDXButton>
+            </Button>
         </StyledCard>
     </CardGroup>
 </ContentSection>
