+++ content/services/classroom.mdx
@@ -21,12 +21,12 @@ CCV provides access to HPC resources for classes, workshops, demonstrations, and
 * Oscar is a shared resource, so access nodes and speed cannot be guaranteed.
 
 <ButtonGroup>
-  <MDXButton href="https://docs.ccv.brown.edu/oscar/account-types">
+  <Button href="https://docs.ccv.brown.edu/oscar/account-types">
     Learn More
-  </MDXButton>
-  <MDXButton href="mailto:support@ccv.brown.edu">
+  </Button>
+  <Button href="mailto:support@ccv.brown.edu">
     Request Support
-  </MDXButton>
+  </Button>
 </ButtonGroup>
 </ContentSection>
 
@@ -38,20 +38,20 @@ Computational Notebooks provide a convenient, cloud-hosted way to serve Jupyter 
 
 Google Colab comes bundled with most Python-specific software libraries, and it supports real-time collaboration. It integrates with Google Drive and should be sufficient for most classroom or workshop needs.
 
-<MDXButton href="https://colab.google">
+<Button href="https://colab.google">
   Learn More
-</MDXButton>
+</Button>
 
 ## Brown JupyterLab
 
 For more advanced needs, Brown's JupyterHub may be a good fit for your needs. If you are an instructor, CCV can provide access to JupyterHub for your class or workshop, and Digital Learning and Design (DLD) can assist with integrating computational assignments into curricula. The implementation is supported by Brown OIT; please follow the link below to request an instance for your class. OIT staff will respond to your request to begin the setup process. We ask that requests for JupyterHub be made at least two months in advance of expected course deployment.
 
 <ButtonGroup>
-  <MDXButton href="https://docs.ccv.brown.edu/jupyterhub">
+  <Button href="https://docs.ccv.brown.edu/jupyterhub">
     Documentation
-  </MDXButton>
-  <MDXButton href="https://docs.google.com/forms/d/e/1FAIpQLSct9rFCxLhPIezHI-RYRyEuSnvHrPZLMuUSFRTriIyd_3TAfA/viewform?usp=sf_link">
+  </Button>
+  <Button href="https://docs.google.com/forms/d/e/1FAIpQLSct9rFCxLhPIezHI-RYRyEuSnvHrPZLMuUSFRTriIyd_3TAfA/viewform?usp=sf_link">
     Request a Hub
-  </MDXButton>
+  </Button>
 </ButtonGroup>
 </ContentSection>
