+++ content/services/oscar.mdx
@@ -7,9 +7,9 @@ description: |
 
 The high-performance computing (HPC) resources at CCV equip the Brown research community with the tools they need to perform complex numerical simulations, modeling, and data analysis. Oscar, CCV's primary research computing cluster, consists of more than 300 multi-core nodes sharing a high-performance interconnect and file system. For more technical details, please see the system overview.
 
-<MDXButton href="https://docs.ccv.brown.edu/oscar/system-overview">
+<Button href="https://docs.ccv.brown.edu/oscar/system-overview">
   View System Hardware
-</MDXButton>
+</Button>
 
 </ContentSection>
 
@@ -29,7 +29,7 @@ A large collection of software is available on Oscar, and CCV staff can help acq
   - Paraview
   - ...and more
 
-<MDXButton href="https://brown.co1.qualtrics.com/jfe/form/SV_0GtBE8kWJpmeG4B">
+<Button href="https://brown.co1.qualtrics.com/jfe/form/SV_0GtBE8kWJpmeG4B">
   Request an Account
-</MDXButton>
+</Button>
 </ContentSection>
