+++ content/services/storage.mdx
@@ -30,12 +30,12 @@ Google also has a really nice feature where you can scan in handwritten document
 - Limitations: Data transfer speeds may be very limited, Globus can provide high bandwidth data transfers.
 
 <ButtonGroup>
-  <MDXButton href="https://ithelp.brown.edu/kb/48-google-drive">
+  <Button href="https://ithelp.brown.edu/kb/48-google-drive">
     Documentation
-  </MDXButton>
-  <MDXButton href="https://workspace.google.com/learn-more/security/security-whitepaper/page-1/">
+  </Button>
+  <Button href="https://workspace.google.com/learn-more/security/security-whitepaper/page-1/">
     Security
-  </MDXButton>
+  </Button>
 </ButtonGroup>
 </ContentSection>
 
@@ -43,9 +43,9 @@ Google also has a really nice feature where you can scan in handwritten document
 
 [Globus](https://www.globus.org/data-transfer) is a secure, reliable, research data management service. With Globus, subscribers can move, share, and discover data via a single interface–whether your files live on a supercomputer, lab cluster, tape archive, public cloud or your laptop–you can manage this data from anywhere via just a web browser and using your existing identities. Transfers are facilitated through endpoints. Endpoints are data locations pointing to subscribers' specific datasets like a researcher's personal Google drive, research share or a folder on a computer.
 
-<MDXButton href="https://docs.ccv.brown.edu/globus">
+<Button href="https://docs.ccv.brown.edu/globus">
   Globus
-</MDXButton>
+</Button>
 </ContentSection>
 
 <ContentSection title="Campus File Storage / Department File Services">
@@ -55,9 +55,9 @@ Departmental File Services provides University departments with a location in wh
 - Best for: Backing up and sharing official department documents, ensuring the longevity of documents after the file authors leave Brown.
 - Limitations: Can only be accessed on the Brown network or with VPN. Sharing is done via Brown AD groups; no sharing with people outside of Brown.
 
-<MDXButton href="https://ithelp.brown.edu/kb/37-department-file-services">
+<Button href="https://ithelp.brown.edu/kb/37-department-file-services">
   Support Documentation
-</MDXButton>
+</Button>
 </ContentSection>
 
 <ContentSection title="Campus File Storage Non-Replicated / Research Services">
@@ -96,9 +96,9 @@ The Brown University Library maintains the repository as a service to the Brown 
 
 Faculty and researchers interested in using the Brown Digital Repository as a platform forprogrammatic data management, storage, and publication should contact the Library (bdr@brown.edu) for information about opportunities for research consulting and project development support.
 
-<MDXButton href="https://repository.library.brown.edu/studio/about/">
+<Button href="https://repository.library.brown.edu/studio/about/">
   Documentation
-</MDXButton>
+</Button>
 </ContentSection>
 
 <ContentSection title="LabArchives">
@@ -109,9 +109,9 @@ LabArchives at Brown provides unlimited storage space. The current size limit pe
 
 LabArchives at Brown is not approved for storing files containing Personally Identifiable Information (PII), Protected Health Information (PHI), or Brown Restricted Information.
 
-<MDXButton href="https://library.brown.edu/info/labarchives/">
+<Button href="https://library.brown.edu/info/labarchives/">
   Documentation
-</MDXButton>
+</Button>
 </ContentSection>
 
 <ContentSection title="Oscar Storage">
@@ -124,12 +124,12 @@ Oscar storage, also known as Computational Data Storage, is a high-performance d
 - Rate: $100/TB/Year when storing above free/grant allocations
 
 <ButtonGroup>
-  <MDXButton href="https://ccv.brown.edu/services/computing#high-performance-computing-(oscar)">
+  <Button href="https://ccv.brown.edu/services/computing#high-performance-computing-(oscar)">
     Documentation
-  </MDXButton>
-  <MDXButton href="https://brown.co1.qualtrics.com/jfe/form/SV_0GtBE8kWJpmeG4B">
+  </Button>
+  <Button href="https://brown.co1.qualtrics.com/jfe/form/SV_0GtBE8kWJpmeG4B">
     Request Service
-  </MDXButton>
+  </Button>
 </ButtonGroup>
 </ContentSection>
 
@@ -140,9 +140,9 @@ Oscar storage, also known as Computational Data Storage, is a high-performance d
 - Best for: Storing data with data usage agreements, FISMA, etc.
 - Rate: $100/TB/Year when storing above free/grant allocations
 
-<MDXButton href="https://www.brown.edu/cis/forms/Stronghold/shold.php">
+<Button href="https://www.brown.edu/cis/forms/Stronghold/shold.php">
   Request Service
-</MDXButton>
+</Button>
 </ContentSection>
 
 <ContentSection title="Hibernate">
@@ -155,7 +155,7 @@ Oscar storage, also known as Computational Data Storage, is a high-performance d
   * Hot Spots (Growth, user(s)/Group(s))
   * Action: Archive, Delete, Recover
 
-<MDXButton href="https://brown.atlassian.net/servicedesk/customer/portal/16/user/login?destination=portal%2F16%2Fgroup%2F55%2Fcreate%2F468">
+<Button href="https://brown.atlassian.net/servicedesk/customer/portal/16/user/login?destination=portal%2F16%2Fgroup%2F55%2Fcreate%2F468">
   Request Hibernate Service
-</MDXButton>
+</Button>
 </ContentSection>
