+++ content/services/stronghold.mdx
@@ -9,9 +9,9 @@ Stronghold has been established by Brown University to be compatible with Federa
 
 Each Principal Investigator (PI) is given a dedicated environment for their project to support their researchers, graduate students, and collaborators. View the Stronghold documentation and learn more about the offered features below.
 
-<MDXButton href="https://docs.ccv.brown.edu/stronghold/">
+<Button href="https://docs.ccv.brown.edu/stronghold/">
   View Documentation
-</MDXButton>
+</Button>
 
 ## Key Features
 
@@ -31,11 +31,11 @@ Each Principal Investigator (PI) is given a dedicated environment for their proj
   Dual Firewall - Provides added layer of security for traffic coming in and out of your system
 
 <ButtonGroup>
-  <MDXButton href="https://brown.atlassian.net/servicedesk/customer/portal/22/group/62/create/256">
+  <Button href="https://brown.atlassian.net/servicedesk/customer/portal/22/group/62/create/256">
     Request an Environment
-  </MDXButton>
-  <MDXButton href="https://it.brown.edu/computing-policies/risk-classifications">
+  </Button>
+  <Button href="https://it.brown.edu/computing-policies/risk-classifications">
     OIT Risk Classification
-  </MDXButton>
+  </Button>
 </ButtonGroup>
 </ContentSection>
