+++ content/services/user-support.mdx
@@ -22,7 +22,7 @@ The Advanced Research IT Solutions group offers specialized IT expertise to acad
 - Managed Database Hosting
 - Consultancy-based Research Software Engineering
 
-<MDXButton href="mailto:ccv-res-sys@brown.edu">
+<Button href="mailto:ccv-res-sys@brown.edu">
   Contact Us
-</MDXButton>
+</Button>
 </ContentSection>
